+++ cmd/tg/hook_audit_lock_posix.go
@@ -1,0 +1,24 @@
+//go:build aix || darwin || dragonfly || freebsd || linux || netbsd || openbsd || solaris
+
+package main
+
+import (
+	"os"
+
+	"golang.org/x/sys/unix"
+)
+
+func tryAuditFileLock(f *os.File) (bool, error) {
+	err := unix.Flock(int(f.Fd()), unix.LOCK_EX|unix.LOCK_NB)
+	if err == nil {
+		return true, nil
+	}
+	if err == unix.EWOULDBLOCK || err == unix.EAGAIN {
+		return false, nil
+	}
+	return false, err
+}
+
+func unlockAuditFile(f *os.File) error {
+	return unix.Flock(int(f.Fd()), unix.LOCK_UN)
+}